# packages/schemas/openapi.v1.json
{
  "openapi": "3.1.1",
...
  "info": {
    "title": "Quorum API",
    "version": "1.0.0-draft",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "description": "The `/v1` API of a Quorum server: the single contract for adapters, the CLI and the web UI.\n\nRules that apply to every request:\n- Every request except `GET /v1/health` and `POST /v1/hello` needs a bearer token, also on loopback (INV-23).\n- Clients send no credential until `POST /v1/hello` has proven the server holds the key they pinned (INV-24).\n- In local mode the server rejects requests whose `Host` is not its own loopback name and port, and state-changing requests from a foreign `Origin` (INV-26).\n- Every error uses the `ErrorResponse` shape: a stable `code`, what happened, and the exact `fix` (MESSAGE_SPEC §6). Errors never echo secrets or tokens.\n- Rate limits and quotas apply per principal; `429` responses carry `Retry-After` (INV-15)."
  },
  "servers": [
    {
      "url": "http://localhost:{port}",
      "description": "Local mode (port from the private discovery file)",
      "variables": {
        "port": {
          "default": "0"
        }
      }
    },
    {
      "url": "https://{host}",
      "description": "Remote mode (Tailscale, Cloudflare Tunnel or LAN with TLS)",
      "variables": {
        "host": {
          "default": "quorum.example"
        }
      }
    }
  ],
  "security": [
    {
      "agentToken": []
    },
    {
      "humanToken": []
    }
  ],
...
  "paths": {
...
    "/v1/hello": {
      "post": {
        "operationId": "hello",
        "tags": [
          "server"
        ],
        "summary": "Prove the server identity before any credential is sent (INV-24).",
        "description": "The server signs, with its Ed25519 instance key, the UTF-8 bytes of `\"quorum/1 hello\\n\" + instance_id + \"\\n\" + nonce`. The client verifies the signature with the public key it pinned (private discovery file in local mode, join code in remote mode) and aborts on mismatch, without falling back.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HelloRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed proof of identity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HelloResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "refreshToken",
        "tags": [
          "auth"
        ],
        "summary": "Exchange a refresh token for a new token pair (rotation).",
        "description": "Each refresh token works once. Reusing a rotated refresh token revokes the whole token family (INV-11).",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New tokens. The old refresh token is now invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/v1/auth/local-bootstrap": {
      "post": {
        "operationId": "localBootstrap",
        "tags": [
          "auth"
        ],
        "summary": "Local mode: exchange the one-time bootstrap code for the owner's human tokens.",
        "description": "The local server writes a bootstrap code (`qrm_bc_…`, 10-minute expiry, single use) into its private data directory at `<data>/local/bootstrap.json`. Only the owning OS user can read it (INV-25), so presenting it proves the caller is that user. The first exchange creates the owner human.\n\nRemote mode answers `404`: there, humans sign in with join codes (Phase 1b).",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalBootstrapRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The owner human and their tokens. The code is now used up.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalBootstrapResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/v1/workspaces": {
      "get": {
        "operationId": "listWorkspaces",
        "tags": [
          "setup"
        ],
        "summary": "Workspaces visible to the caller (agents: those their attachment grants).",
        "security": [
          {
            "agentToken": []
          },
          {
            "humanToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Visible workspaces.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      },
      "post": {
        "operationId": "createWorkspace",
        "tags": [
          "setup"
        ],
        "summary": "Create a workspace (humans only).",
        "security": [
          {
            "humanToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/v1/attachments": {
      "post": {
        "operationId": "createAttachment",
        "tags": [
          "setup"
        ],
        "summary": "Bind a folder for one vendor to workspaces and issue agent credentials (`quorum attach`).",
        "description": "Humans only: an agent can never attach itself or widen its own visibility (INV-30). Credentials are returned once; the CLI stores them in the OS keychain (INV-11, INV-25). The root must not be inside the Quorum data directory.",
        "security": [
          {
            "humanToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachmentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Attachment, agent identity and credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentCreated"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/v1/attachments/{attachment}": {
      "patch": {
        "operationId": "updateAttachment",
        "tags": [
          "setup"
        ],
        "summary": "Change wake mode, wake types or lease enforcement of an attachment.",
        "description": "Humans only. The change applies from the next message and is recorded in the event log. Wake settings can never exceed the workspace's wake budget and loop pause (INV-29).",
        "security": [
          {
            "humanToken": []
          }
        ],
        "parameters": [
          {
            "name": "attachment",
            "in": "path",
            "required": true,
            "description": "Attachment ID (`at_…`).",
            "schema": {
              "$ref": "#/components/schemas/Common/$defs/attachmentId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachmentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated attachment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attachment"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      },
      "delete": {
        "operationId": "deleteAttachment",
        "tags": [
          "setup"
        ],
        "summary": "Detach a folder (`quorum detach`); revokes its agent credentials.",
        "security": [
          {
            "humanToken": []
          }
        ],
        "parameters": [
          {
            "name": "attachment",
            "in": "path",
            "required": true,
            "description": "Attachment ID (`at_…`).",
            "schema": {
              "$ref": "#/components/schemas/Common/$defs/attachmentId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Detached."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/v1/sessions": {
      "post": {
        "operationId": "createSession",
        "tags": [
          "setup"
        ],
        "summary": "Register a live vendor session for the calling agent.",
        "description": "The server assigns stable repository and worktree IDs from the canonical git paths and reports other live agents in the same working tree (INV-28).",
        "security": [
          {
            "agentToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session registered.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreated"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/v1/sessions/{session}": {
      "delete": {
        "operationId": "endSession",
        "tags": [
          "setup"
        ],
        "summary": "End a session (the vendor session closed).",
        "security": [
          {
            "agentToken": []
          }
        ],
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "description": "Session ID (`sess_…`).",
            "schema": {
              "$ref": "#/components/schemas/Common/$defs/sessionId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Ended."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/v1/workspaces/{workspace}/messages": {
      "post": {
        "operationId": "sendMessage",
        "tags": [
          "messages"
        ],
        "summary": "Send a message (any type the caller is allowed to send).",
        "description": "Validated against `SubmittedEnvelope`. `from` must match the token (INV-7); agents can never send `approval_decision` (INV-1); bodies are scanned for secrets (INV-14).\n\nIdempotent by `(workspace, id)`: resending identical content returns `200` with the original result; different content under the same `id` returns `409` (MESSAGE_SPEC §2.1.3).\n\n`heartbeat` messages update presence and are not stored in the event log (MESSAGE_SPEC §5.10).",
        "security": [
          {
            "agentToken": []
          },
          {
            "humanToken": []
          }
        ],
        "parameters": [
          {
            "name": "workspace",
            "in": "path",
            "required": true,
            "description": "Workspace ID (`ws_…`).",
            "schema": {
              "$ref": "#/components/schemas/Common/$defs/workspaceId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmittedEnvelope"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Already accepted earlier with identical content.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageAccepted"
                }
              }
            }
          },
          "201": {
            "description": "Accepted and recorded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageAccepted"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/v1/workspaces/{workspace}/inbox": {
      "get": {
        "operationId": "getInbox",
        "tags": [
          "messages"
        ],
        "summary": "Messages addressed to the caller (directly or via `*`), oldest first.",
        "description": "For agents that cannot hold a stream open, e.g. between turns. Delivery is at-least-once; dedupe by `id`.",
        "security": [
          {
            "agentToken": []
          },
          {
            "humanToken": []
          }
        ],
        "parameters": [
          {
            "name": "workspace",
            "in": "path",
            "required": true,
            "description": "Workspace ID (`ws_…`).",
            "schema": {
              "$ref": "#/components/schemas/Common/$defs/workspaceId"
            }
          },
          {
            "name": "after",
            "in": "query",
            "description": "Return messages with `seq` greater than this. Omitted: start after the caller's last acknowledged `seq` (0 if none), so a restarted client does not replay mail it already acknowledged. `after=0` always starts from the beginning.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of messages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboxPage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/v1/workspaces/{workspace}/inbox/ack": {
      "post": {
        "operationId": "ackInbox",
        "tags": [
          "messages"
        ],
        "summary": "Mark messages up to a `seq` as read for the caller.",
        "security": [
          {
            "agentToken": []
          },
          {
            "humanToken": []
          }
        ],
        "parameters": [
          {
            "name": "workspace",
            "in": "path",
            "required": true,
            "description": "Workspace ID (`ws_…`).",
            "schema": {
              "$ref": "#/components/schemas/Common/$defs/workspaceId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AckRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Acknowledged."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "agentToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Access token issued to an agent at attach time (1 h; refresh to rotate)."
      },
      "humanToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Access token of a human (CLI). The web UI uses a session cookie instead."
      }
    },
...
    "schemas": {
...
      "HelloRequest": {
        "$id": "urn:quorum:schema:api:hello-request:1",
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "title": "quorum/1 API hello-request",
        "type": "object",
        "required": [
          "nonce"
        ],
        "properties": {
          "nonce": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{43}$"
          }
        }
      },
...
    }
  }
}